update include/lib files of ffmpeg to 0.5.1 version. Add support Windows64 under mingw-w64.
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@2798 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/include/opencv/cxoperations.hpp
@@ -2581,7 +2581,7 @@ template<typename _Tp> inline ptrdiff_t operator - (const SeqIterator<_Tp>& a,
                                                     const SeqIterator<_Tp>& b)
 {
     ptrdiff_t delta = a.index - b.index, n = a.seq->total;
-    if( std::abs(delta) > n )
+    if( std::abs(static_cast<long>(delta)) > n )
         delta += delta < 0 ? n : -n;
     return delta;
 }
